moves validation to utils folder
